ai_apps/Pages/3_Dashboard_Analyzer.py
# Q&A Chatbot
#from langchain.llms import OpenAI

from dotenv import load_dotenv

load_dotenv()  # take environment variables from .env.

import streamlit as st
import os
import pathlib
import textwrap
from PIL import Image


import google.generativeai as genai


os.getenv("GEMINI_API_KEY")
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

## Function to load OpenAI model and get respones

def get_gemini_response(input,image):
    model = genai.GenerativeModel('gemini-pro-vision')
    if input!="":
       response = model.generate_content([input,image])
    else:
       response = model.generate_content(image)
    return response.text

##initialize our streamlit app

st.set_page_config(
    page_title="Dashboard Analyzer",
    page_icon=":orange_heart:",
)
st.title("Dashboard Analyzer")

st.header("Dashboard Analyzer Application")
input='''I want to know the dashboard details of All the metrics in table format. 
        The table should have four columns that is matrics name, matrix value , status, Remedial Actions. 
        Depending on each matrix create a seperate table outling all the important information related to that matrix'''
uploaded_file = st.file_uploader("Choose an image...", type=["jpg", "jpeg", "png"])
image=""   
if uploaded_file is not None:
    image = Image.open(uploaded_file)
    st.image(image, caption="Uploaded Image.", use_column_width=True)


submit=st.button("Analyze the dashboard")

## If ask button is clicked

if submit:
    
    response=get_gemini_response(input,image)
    st.subheader("The Response is")
    st.write(response)